+++ src/Features/Users/userSlice.js
@@ -1,29 +1,58 @@
-import {
-    createSlice
-} from '@reduxjs/toolkit'
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
+import { FacebookAuthProvider, signInWithPopup } from 'firebase/auth';
+import { auth } from '../../firebase-config';
 
 const initialState = {
-    user: null
-}
+  user: {},
+  status: 'idle',
+  error: null,
+};
+export const signInWithFacebook = createAsyncThunk(
+  'user/signInWithFacebook',
+  async (payload, { rejectWithValue }) => {
+    const provider = new FacebookAuthProvider();
+    try {
+      const { user } = await signInWithPopup(auth, provider);
+      payload();
+      return JSON.stringify(user);
+    } catch (error) {
+      return rejectWithValue(JSON.stringify(error));
+    }
+  }
+);
 
 const userSlice = createSlice({
-    name: 'user',
-    initialState,
-    reducers: {
-        login: (state, action) => {
-            state.user = action.payload;
-        },
-        logout: (state) => {
-            state.user = null;
-        },
+  name: 'user',
+  initialState,
+  reducers: {
+    login: (state, action) => {
+      state.user = action.payload;
     },
-})
+    logout: (state) => {
+      state.user = null;
+    },
+  },
+  extraReducers: (builder) => {
+    builder
+      // Login with Facebook
+      .addCase(signInWithFacebook.pending, (state) => {
+        state.status = 'loading';
+      })
+      .addCase(signInWithFacebook.fulfilled, (state, { payload }) => {
+        state.status = 'succeeded';
+        state.user = JSON.parse(payload);
+      })
+      .addCase(signInWithFacebook.rejected, (state, { payload }) => {
+        state.status = 'failed';
+        state.error = JSON.parse(payload);
+      });
+  },
+});
 
-export const {
-    login,
-    logout
-} = userSlice.actions;
+export const { login, logout } = userSlice.actions;
 
-export const selectUser = (state) => state.user.user;
+export const user = (state) => state.user;
+export const status = (state) => state.status;
+export const error = (state) => state.error;
 
-export default userSlice
+export default userSlice;
