+++ src/app/store.js
@@ -1,10 +1,7 @@
 import {
   configureStore
 } from '@reduxjs/toolkit';
-import userReducer from '../Features/Users/userSlice';
-
-import { configureStore } from '@reduxjs/toolkit';
-import userReducer from '../features/user/userSlice';
+import userReducer from '../features/Users/userSlice';
 
 export const store = configureStore({
   reducer: {
