+++ src/components/SignUp/SignUp.test.js
@@ -1,7 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { Provider } from 'react-redux';
 import SignUp from './SignUp';
 import { store } from '../../app/store';
